+++ torch/_export/utils.py
@@ -1,11 +1,11 @@
 import dataclasses
-
+import math
 from typing import Any, Dict, Iterable, List, Optional, Tuple, Type
 
 import torch
 
 from torch._export import ExportedProgram
-
+from torch._subclasses.fake_tensor import FakeTensor
 from torch.utils._pytree import (
     _register_pytree_node,
     Context,
@@ -25,37 +25,87 @@ SERIALIZED_DATACLASS_TO_PYTHON_DATACLASS: Dict[str, Type[Any]] = {}
 def _check_input_constraints_pre_hook(self, *args, **kwargs):
     flat_args, _ = tree_flatten(args)
     return _check_input_constraints_for_graph(
-        self.graph,
-        range_constraints=self.range_constraints,
-        equality_constraints=self.equality_constraints,
-    )(*flat_args)
+        [node for node in self.graph.nodes if node.op == "placeholder"],
+        flat_args,
+        self.range_constraints,
+    )
 
 
 def _check_input_constraints_for_graph(
-    graph: torch.fx.Graph, range_constraints, equality_constraints
+    input_placeholders: List[torch.fx.Node], args, range_constraints
 ):
+    def check(cond, msg):
+        if not cond:
+            # TODO(avik): maybe add more context, e.g., graph signature
+            raise RuntimeError(msg)
+
+    import sympy
+
     from torch._export.passes.add_runtime_assertions_for_constraints_pass import (
-        _AddRuntimeAssertionsForConstraintsPass,
+        _convert_range_to_int,
     )
 
-    def inner(*args):
-        # TODO(zhxchen17) Don't generate a runtime graph on the fly.
-        _assertion_graph = torch.fx.GraphModule({}, torch.fx.Graph())
-        for p in graph.nodes:
-            if p.op != "placeholder":
-                continue
-            new_p = _assertion_graph.graph.placeholder(p.name)
-            new_p.meta = p.meta
-        _assertion_graph.graph.output(())
-        _assertion_graph_res = _AddRuntimeAssertionsForConstraintsPass(
-            range_constraints,
-            equality_constraints,
-        )(_assertion_graph)
-        assert _assertion_graph_res is not None
-        _assertion_graph = _assertion_graph_res.graph_module
-        _assertion_graph(*args)
+    check(
+        len(args) == len(input_placeholders),
+        "Unexpected number of inputs "
+        f"(expected {len(input_placeholders)}, got {len(args)})",
+    )
+    # NOTE: export already guarantees that the same symbol is used in metadata
+    # for all InputDims related by equality constraints, so we can just unify
+    # symbols with given input dimension values to check equality constraints.
+    unification_map: "Dict[sympy.Symbol, Any]" = {}
+    for arg, node in zip(args, input_placeholders):
+        node_val = node.meta["val"]
+        if isinstance(node_val, FakeTensor):
+            check(
+                isinstance(arg, torch.Tensor),
+                f"Expected input {node.name} to be a tensor, but got {type(arg)}",
+            )
+            check(
+                len(node_val.shape) == len(arg.shape),
+                f"Unexpected number of dimensions in input {node.name}.shape "
+                f"(expected {node_val.shape}, got {arg.shape})",
+            )
+            for j, (arg_dim, node_dim) in enumerate(zip(arg.shape, node_val.shape)):
+                if isinstance(node_dim, torch.SymInt):
+                    if node_dim.node.expr in unification_map:
+                        existing_dim = unification_map[node_dim.node.expr]
+                        check(
+                            arg_dim == existing_dim,
+                            f"Expected input {node.name}.shape[{j}] to be equal to "
+                            f"{existing_dim}, but got {arg_dim}",
+                        )
+                    else:
+                        unification_map[node_dim.node.expr] = arg_dim
 
-    return inner
+                    if node_dim.node.expr in range_constraints:
+                        min_val, max_val = _convert_range_to_int(
+                            range_constraints[node_dim.node.expr]
+                        )
+                        # NOTE: we allow dimensions to be 0/1 at runtime
+                        if min_val > 2:
+                            check(
+                                arg_dim >= min_val,
+                                f"Expected input {node.name}.shape[{j}] to be >= "
+                                f"{min_val}, but got {arg_dim}",
+                            )
+                        if max_val < math.inf:
+                            check(
+                                arg_dim <= max_val,
+                                f"Expected input {node.name}.shape[{j}] to be <= "
+                                f"{max_val}, but got {arg_dim}",
+                            )
+                else:
+                    check(
+                        arg_dim == node_dim,
+                        f"Expected input {node.name}.shape[{j}] to be equal to "
+                        f"{node_dim}, but got {arg_dim}",
+                    )
+        elif isinstance(node_val, (int, float, str)):
+            check(
+                type(arg) == type(node_val) and arg == node_val,
+                f"Expected input {node.name} to be equal to {node_val}, but got {arg}",
+            )
 
 
 def register_dataclass_as_pytree_node(
